+++ PanaceaServices/PanaceaServicesSoapClient/pom.xml
@@ -10,12 +10,12 @@
         <maven.compiler.source>1.8</maven.compiler.source>
         <maven.compiler.target>1.8</maven.compiler.target>
     </properties>
-    <repositories>
+<!--    <repositories>
         <repository>
             <id>My Grid</id>
             <url>http://www.mygrid.org.uk/maven/repository/</url>
         </repository>
-    </repositories>
+    </repositories>-->
     <dependencies>
         <dependency>
             <groupId>junit</groupId>
@@ -48,11 +48,11 @@
             <artifactId>axis</artifactId>
             <version>1.4</version>
         </dependency>
-        <dependency>
+<!--        <dependency>
             <groupId>org.soaplab</groupId>
             <artifactId>taverna-soaplab-client</artifactId>
             <version>2.2.0</version>
-        </dependency>
+        </dependency>-->
         <!-- https://mvnrepository.com/artifact/commons-validator/commons-validator -->
         <dependency>
             <groupId>commons-validator</groupId>
@@ -61,17 +61,12 @@
         </dependency>
         <dependency>
             <groupId>${project.groupId}</groupId>
-            <artifactId>IlcFillSimpleTypes</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>${project.groupId}</groupId>
-            <artifactId>IlcUtils</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>${project.groupId}</groupId>
             <artifactId>IlcConsumerAndProvider</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>PanaceaServiceImpl</artifactId>
             <version>${project.version}</version>
         </dependency>
     </dependencies>
@@ -112,5 +107,5 @@
             </plugin>
         </plugins>
     </build>
-    <name>panaceaServicesSoapClient</name>
+    <name>PanaceaServicesSoapClient</name>
 </project>
