+++ PanaceaServices/PanaceaServicesSoapClient/src/main/java/it/cnr/ilc/soapclient/app/Theservice.java
@@ -6,7 +6,8 @@
 package it.cnr.ilc.soapclient.app;
 
 import it.cnr.ilc.ilcutils.Vars;
-import it.cnr.ilc.soapclient.i.PanaceaService;
+import it.cnr.ilc.panacea.service.i.PanaceaService;
+
 import java.util.Map;
 import java.util.logging.Level;
 import java.util.logging.Logger;
